+++ R/wrf_create.R
@@ -1,29 +1,39 @@
-#' Create emission files to the WRF-Chem
+#' Create emission files for the WRF-Chem model
 #'
-#' @description Create an emission file from wrfinput
+#' @description Create WRF-chem emission files using information from the WRF initial condictions (wrfinput) file(s). The wrfinput file of the corresponding domain is read from the current folder or from the wrfinput_dir.
 #'
-#' @param wrfinput_dir folder with the wrfinput file(s)
+#' There are two emission styles available: the 12 hour pair of emissions (that will be recycled by the model) using io_style_emissions = 1 and the date_hour format using io_style_emissions = 2 (default).
+#'
+#' The initial time is the original (wrfinput file) adjusted by the day_offset argument, this argument can be useful for split the emissions into several files or for a restarted simulation. The emissions are recorded at the interval of 60 minutes (or the auxinput5_interval_m argument) for 1 time (or frames_per_auxinput5 argument times).
+#'
+#' The variables created on output file is based on emis_opt data or a character vector contains the species, any change in variables need to be followed by a change in the n_aero for the correspondent number of aerosol species in the emission file (the n_aero last variables).
+#'
+#' Title argument will be written on global attribute TITLE, from the version 4.0 the model checks if the TITLE version contains "V4.", this can be disabled setting 'force_use_old_data = .true.' on WRF namelist.input.
+#'
+#' @param wrfinput_dir input folder with the wrfinput file(s)
 #' @param wrfchemi_dir output folder
-#' @param domains domain or domains to process
+#' @param domains domain / domains to be process
 #' @param frames_per_auxinput5 value from wrf &time_control namelist.input,
-#' number of times in a single emission file
+#' number of times (frames) in a single emission file
 #' @param auxinput5_interval_m value from wrf &time_control namelist.input,
-#' interval in minutes between different times
-#' @param day_offset number of days (can be a fracion) to create multiple files
-#' @param io_style_emissions from wrf &chem namelist.input
-#' @param kemit from wrf &chem namelist.input, number of levels of the emission file
-#' @param variaveis emission species, can be used data(emis_opt)
+#' interval in minutes between different times (frames) see Details
+#' @param day_offset number of days (can be a fraction) see Details
+#' @param io_style_emissions from wrf &chem namelist.input see Details
+#' @param kemit from wrf &chem namelist.input, number of vertical levels of the emission file
+#' @param variables emission species, can be used data(emis_opt)
 #' @param n_aero number of aerosol species
 #' @param COMPRESS integer between 1 (least compr) and 9 (most compr) or NA for
 #' no compression
 #' @param force_ncdf4 force NetCDF4 format
+#' @param title TITLE attribute for the NetCDF
 #' @param verbose print file info
-#' @note to use io_style_emissions = 1, use day_offset increased by 0.5
-#' (to increase 12h)
 #'
-#' @note Windowns users need to rename the emission files
+#' @note Using io_style_emissions = 1, the wrfchemi_00z will be generated with day_offset = 0 and
+#' wrfchemi_12z with day_offset = 0.5 (frames_per_auxinput5 and auxinput5_interval_m will have no effect).
+#'
+#' @note Windows users need to rename the emission files
 #' from 'wrfchemi_d01_2011-08-01_00_00_00' to 'wrfchemi_d01_2011-08-01_00:00:00'
-#' before run wrf.exe with these files
+#' or change in namelist the defoult filename before run wrf.exe with these files.
 #'
 #' @author Daniel Schuch
 #'
@@ -40,80 +50,65 @@
 #'
 #' dir.create(file.path(tempdir(), "EMISS"))
 #'
+#' # emissions on date_hour style
 #' wrf_create(wrfinput_dir         = system.file("extdata", package = "eixport"),
 #'            wrfchemi_dir         = file.path(tempdir(), "EMISS"),
 #'            domains              = 1:2,
-#'            frames_per_auxinput5 = 24,
+#'            frames_per_auxinput5 = 25,
 #'            auxinput5_interval_m = 60,
-#'            day_offset           = 0,
 #'            verbose              = TRUE)
 #'
-#' # emission for the last timestep
-#'
+#' # emissions on 00z / 12z style, create the 00z
 #' wrf_create(wrfinput_dir         = system.file("extdata", package = "eixport"),
 #'            wrfchemi_dir         = file.path(tempdir(), "EMISS"),
 #'            domains              = 1:2,
-#'            frames_per_auxinput5 = 1,
-#'            auxinput5_interval_m = 60,
-#'            day_offset           = 1,
+#'            io_style_emissions   = 1,
+#'            day_offset           = 0,
+#'            verbose              = TRUE,
+#'            )
+#' # emissions on 00z / 12z style, create the 12z
+#' wrf_create(wrfinput_dir         = system.file("extdata", package = "eixport"),
+#'            wrfchemi_dir         = file.path(tempdir(), "EMISS"),
+#'            domains              = 1:2,
+#'            io_style_emissions   = 1,
+#'            day_offset           = 0.5,
 #'            verbose              = TRUE)
 #'}
 
-wrf_create  <- function(wrfinput_dir         = "",
-                        wrfchemi_dir         = "",
+wrf_create  <- function(wrfinput_dir         = getwd(),
+                        wrfchemi_dir         = getwd(),
                         domains              = 1,
                         frames_per_auxinput5 = 1,
                         auxinput5_interval_m = 60,
                         day_offset           = 0,
                         io_style_emissions   = 2,
                         kemit                = 1,
-                        variaveis = c("E_NO2",
-                                      "E_NO",
-                                      "E_TOL",
-                                      "E_XYL",
-                                      "E_ALD",
-                                      "E_ALDX",
-                                      "E_SO2",
-                                      "E_CO",
-                                      "E_OLT",
-                                      "E_OLI",
-                                      "E_OL2",
-                                      "E_NH3",
-                                      "E_ISO",
-                                      "E_HCL",
-                                      "E_HCHO",
-                                      "E_ETH",
-                                      "E_CH3OH",
-                                      "E_C2H5OH",
-                                      "E_HC3",
-                                      "E_HC5",
-                                      "E_HC8",
-                                      "E_KET",
-                                      "E_ORA2",
-                                      "E_CSL",
-                                      "E_TERP",
-                                      "E_PM25I",
-                                      "E_PM25J",
-                                      "E_ECI",
-                                      "E_ECJ",
-                                      "E_ORGI",
-                                      "E_ORGJ",
-                                      "E_SO4I",
-                                      "E_SO4J",
-                                      "E_NO3J",
-                                      "E_NO3I",
-                                      "E_SO4C",
-                                      "E_NO3C",
-                                      "E_ORGC",
-                                      "E_ECC",
-                                      "E_PM10"),
-                        n_aero               = 15,
+                        variables            = "ecbmz_mosaic",
+                        n_aero               = 14,
                         COMPRESS             = NA,
                         force_ncdf4          = FALSE,
+                        title                = "Anthropogenic emissions for WRF V4.0",
                         verbose              = FALSE)
 {
   a <- Sys.info()["sysname"]
-  if(a[[1]] == "Windows") linux = F else linux = T # to avoid special chacacteres in the filename
+  # to avoid special chacacteres in the filename
+  if(a[[1]] == "Windows") linux = F else linux = T # nocov
+  if(a[[1]] == "Windows")
+    if(io_style_emissions == 2) #nocov
+      cat("\nNOTE: rename the output or change the default filename in namelist before run\n")#nocov
+
+  if(length(variables) == 1){
+    emis_opt <- NULL
+    load(system.file("data/emis_opt.rda", package = "eixport"))
+    if(variables %in% names(emis_opt)){
+      variables <- emis_opt[[variables]]
+    }else{
+      cat(paste(variables,"is not valid, use one of:\n"))           # nocov
+      cat(names(emis_opt))                                          # nocov
+      stop("name, numeric value or a set of variable names")        # nocov
+    }
+
+  }
 
   for(domain in domains){
     # basic information from wrfinput
@@ -124,6 +119,13 @@ wrf_create  <- function(wrfinput_dir         = "",
     date <- as.character(paste(substr(input_time, 1, 10),
                                substr(input_time, 12, 19)))
     date <- as.POSIXct(strptime(date, "%Y-%m-%d %H:%M:%S"))
+
+    if(io_style_emissions == 1){
+      if(day_offset != 0){                    # nocov
+        day_offset <- 0.5                     # nocov
+      }
+    }
+
     date <- date + 86400 * day_offset
     hora         <- format(date,"%H")
     minuto       <- format(date,"%M")
@@ -133,30 +135,29 @@ wrf_create  <- function(wrfinput_dir         = "",
     segundo      <- paste(formatC(segundo, width = 2, format = "d", flag = "0"))
 
     if(io_style_emissions == 1){
-      frames_per_auxinput5 <- 12
-      if(is.integer(day_offset)){
+      frames_per_auxinput5 <- 12  # nocov start
+      if(day_offset == 0){
         h <- "00z"
       }else{
         h <- "12z"
       }
-      file_name <- paste(wrfchemi_dir, "/wrfchemi_", h, "_", "d0",
-                         domain, sep = "")
+      file_name <- paste(wrfchemi_dir, "/wrfchemi_", h, "_", "d0",domain, sep = "") # nocov end
     }
-    if(io_style_emissions ==2){
+    if(io_style_emissions ==2){                                                     # nocov start
       if(linux){
         file_name <- paste(wrfchemi_dir, "/wrfchemi_d0", domain, "_",
                            format(date,"%Y-%m-%d"),
                                    "_",hora,":", minuto,":", segundo, sep = "")
       } else  file_name <- paste(wrfchemi_dir, "/wrfchemi_d0", domain, "_",
                                  format(date,"%Y-%m-%d"),
-                                 "_", hora, "_", minuto, "_", segundo, sep = "")
+                                 "_", hora, "_", minuto, "_", segundo, sep = "")   # nocov end
     }
 
     if(frames_per_auxinput5 == 1){
       file_time  <- paste(format(date,"%Y-%m-%d"),
                           "_", hora, ":", minuto, ":", segundo, sep = "")
     }
-    else{
+    else{                                                                   # nocov start
       file_time    <- character(frames_per_auxinput5)
       for(i in 1:frames_per_auxinput5){
         file_time[i] <- paste(format(date,"%Y-%m-%d"),
@@ -177,17 +178,18 @@ wrf_create  <- function(wrfinput_dir         = "",
                                       width = 2,
                                       format = "d",
                                       flag = "0"))
-      }
+      }                                                                      # nocov end
     }
 
     input_lat    <- ncdf4::ncvar_get(wrfinput, "XLAT")
     input_lon    <- ncdf4::ncvar_get(wrfinput, "XLONG")
     g_atributos  <- ncdf4::ncatt_get(wrfinput, 0)
-    g_atributos  <- c( list(Title = "Anthropogenic emissions",
+    g_atributos  <- c( list(TITLE = title,
                             History = paste("created on",
                                             format(Sys.time(),
                                                    "%Y-%m-%d at %H:%M")),
-                            Author = "R package eixport"),
+                            Author = paste0("R package eixport v",
+                                            utils::packageVersion("eixport"))),
                        g_atributos[4:length(g_atributos)])
     # definition of dimensions
     west_east <- ncdf4::ncdim_def("west_east",
@@ -227,10 +229,10 @@ wrf_create  <- function(wrfinput_dir         = "",
                              dim = list(west_east, south_north),
                              prec = "float",
                              compression = COMPRESS)
-    # GAS fase emissions
-    for(i in 1:(length(variaveis) - n_aero)){
-      assign(variaveis[i],
-             ncdf4::ncvar_def(name = variaveis[i],
+    # GAS phase emissions
+    for(i in 1:(length(variables) - n_aero)){
+      assign(variables[i],
+             ncdf4::ncvar_def(name = variables[i],
                               units = "",
                               dim = list(west_east,
                                          south_north,
@@ -239,10 +241,10 @@ wrf_create  <- function(wrfinput_dir         = "",
                               prec="float",
                               compression = COMPRESS))
     }
-    # AEROSOL emissions
-    for(i in (1+length(variaveis) - n_aero):length(variaveis)){
-      assign(variaveis[i],
-             ncdf4::ncvar_def(name = variaveis[i],
+    # AEROSOLS emissions
+    for(i in (1+length(variables) - n_aero):length(variables)){
+      assign(variables[i],
+             ncdf4::ncvar_def(name = variables[i],
                               units = "",
                               dim = list(west_east,
                                          south_north,
@@ -317,55 +319,55 @@ wrf_create  <- function(wrfinput_dir         = "",
                        south_north$len,
                        emissions_zdim_stag$len,
                        Time$len))
-    # GASES inicializat5ion with zeros
-    for(i in 1:(length(variaveis) - n_aero)){
+    # GASES initialization with zeros
+    for(i in 1:(length(variables) - n_aero)){
       ncdf4::ncvar_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        zero)
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "MemoryOrder",
                        attval = "XYZ")
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "description",
                        attval = "EMISSIONS")
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "units",
                        attval = "mol km^-2 hr^-1")
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "stagger",
                        attval = "Z")
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "FieldType",
                        attval = 104)
     }
-    # AEROSOIS inicializat5ion with zeros
-    for(i in (1 + length(variaveis) - n_aero):length(variaveis)){
+    # AEROSOIS initialization with zeros
+    for(i in (1 + length(variables) - n_aero):length(variables)){
       ncdf4::ncvar_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        zero)
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "MemoryOrder",
                        attval = "XYZ")
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "description",
                        attval = "EMISSIONS")
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "units",
                        attval = "ug m^-2 s^-1")
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "stagger",
                        attval = "Z")
       ncdf4::ncatt_put(emiss_file,
-                       varid = variaveis[i],
+                       varid = variables[i],
                        attname = "FieldType",
                        attval = 104)
     }